core/requirements.txt
beautifulsoup4==4.12.3
Django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.12.1
huggingface-hub==0.20.3
Jinja2==3.1.3
llama-index==0.9.15.post2
Markdown==3.5.2
openai==1.4.0
pillow==10.2.0
pydantic==2.5.2
pydantic_core==2.14.5
pypdf==4.0.2
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.4.0
tiktoken==0.5.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
urllib3==2.2.1
